Move LSPS2 service logic into `liquidity/service/lsps2.rs`
